+++ CHANGELOG.md
@@ -4,7 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]## [0.5.4.0b]
+## [Unreleased]## [0.5.4.1b]
+
+## [0.5.4.0] - 2022-01-11
 ### Changed
 - ScenarioDbManager - Converted text SQL operations to SQLAlchemy operations to support any column-name (i.e. lower, upper, mixed, reserved words)
 - Updated ScenarioDbManager.read_scenario_tables_from_db to selectively read tables from a scenario